requirements.txt

pymongo[srv]
python-dotenv
click==7.1.2
cycler==0.10.0
dnspython==1.16.0
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.5
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
Werkzeug==1.0.1